refactor(graphql): use generated types and improve book service
- **Replaced manual DTO imports**:
  - Used generated `CreateBookInput` and `UpdateBookInput` from `graphql.ts`.
- **Updated ID type**:
  - Changed `id` type from `number` to `string` in `BooksResolver` and `BooksService` to align with `ID!` in the GraphQL schema.
- **Refactored `BooksService`**:
  - Added an in-memory array to store books.
  - Implemented `create()`, `findAll()`, and `findOne()` methods with real functionality.
  - Updated `create()` to generate unique string-based IDs.
