--- src/app.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class AppService {
-    getHello(): string {
-        return 'Hello World!';
-    }
-}
