--- src/books/dto/create-book.input.ts
@@ -1,1 +1,0 @@
-export class CreateBookInput {}
