--- src/books/dto/update-book.input.ts
@@ -1,6 +1,0 @@
-import { CreateBookInput } from './create-book.input';
-import { PartialType } from '@nestjs/mapped-types';
-
-export class UpdateBookInput extends PartialType(CreateBookInput) {
-    id: number;
-}
